Add trakt module
