+++ app/src/main/java/me/banes/chris/tivi/home/discover/DiscoverViewModel.kt
@@ -20,8 +20,6 @@ import android.arch.lifecycle.MutableLiveData
 import io.reactivex.Flowable
 import io.reactivex.functions.BiFunction
 import me.banes.chris.tivi.AppNavigator
-import me.banes.chris.tivi.calls.PopularCall
-import me.banes.chris.tivi.calls.TrendingCall
 import me.banes.chris.tivi.data.Entry
 import me.banes.chris.tivi.data.entities.ListItem
 import me.banes.chris.tivi.data.entities.PopularListItem
@@ -33,6 +31,8 @@ import me.banes.chris.tivi.home.HomeNavigator
 import me.banes.chris.tivi.home.discover.DiscoverViewModel.Section.POPULAR
 import me.banes.chris.tivi.home.discover.DiscoverViewModel.Section.TRENDING
 import me.banes.chris.tivi.trakt.TraktManager
+import me.banes.chris.tivi.trakt.calls.PopularCall
+import me.banes.chris.tivi.trakt.calls.TrendingCall
 import me.banes.chris.tivi.util.AppRxSchedulers
 import timber.log.Timber
 import javax.inject.Inject
