+++ app/src/main/java/me/banes/chris/tivi/home/library/LibraryViewModel.kt
@@ -18,7 +18,6 @@ package me.banes.chris.tivi.home.library
 
 import android.arch.lifecycle.MutableLiveData
 import me.banes.chris.tivi.AppNavigator
-import me.banes.chris.tivi.calls.WatchedCall
 import me.banes.chris.tivi.data.Entry
 import me.banes.chris.tivi.data.entities.ListItem
 import me.banes.chris.tivi.data.entities.TiviShow
@@ -28,6 +27,7 @@ import me.banes.chris.tivi.home.HomeNavigator
 import me.banes.chris.tivi.home.library.LibraryViewModel.Section.WATCHED
 import me.banes.chris.tivi.home.library.LibraryViewModel.Section.WHATS_NEXT
 import me.banes.chris.tivi.trakt.TraktManager
+import me.banes.chris.tivi.trakt.calls.WatchedCall
 import me.banes.chris.tivi.util.AppRxSchedulers
 import timber.log.Timber
 import javax.inject.Inject
