+++ app/src/main/java/me/banes/chris/tivi/home/trending/TrendingShowsViewModel.kt
@@ -16,9 +16,9 @@
 
 package me.banes.chris.tivi.home.trending
 
-import me.banes.chris.tivi.calls.TrendingCall
 import me.banes.chris.tivi.data.entities.TrendingListItem
 import me.banes.chris.tivi.home.HomeNavigator
+import me.banes.chris.tivi.trakt.calls.TrendingCall
 import me.banes.chris.tivi.util.AppRxSchedulers
 import me.banes.chris.tivi.util.EntryViewModel
 import javax.inject.Inject
