+++ base/src/main/java/me/banes/chris/tivi/AppNavigator.kt
@@ -14,18 +14,11 @@
  * limitations under the License.
  */
 
-package me.banes.chris.tivi.calls
+package me.banes.chris.tivi
 
-import android.arch.paging.LivePagedListProvider
-import io.reactivex.Completable
-import io.reactivex.Flowable
+import android.app.PendingIntent
 
-interface Call<in Param, DatabaseOutput> {
-    fun data(): Flowable<DatabaseOutput>
-    fun refresh(param: Param): Completable
-}
-
-interface ListCall<in Param, DatabaseOutput> : Call<Param, List<DatabaseOutput>> {
-    fun liveList(): LivePagedListProvider<Int, DatabaseOutput>
-    val pageSize: Int
+interface AppNavigator {
+    fun provideAuthHandleResponseIntent(requestCode: Int): PendingIntent
+    fun startSettings()
 }
