+++ base/src/main/java/me/banes/chris/tivi/extensions/PreferenceExtensions.kt
