+++ settings.gradle
@@ -16,5 +16,6 @@
  */
 
 include ':base'
+include ':trakt'
 include ':data'
 include ':app'
