+++ trakt/src/main/java/me/banes/chris/tivi/trakt/TraktManager.kt
