+++ trakt/src/main/java/me/banes/chris/tivi/trakt/calls/TrendingCall.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package me.banes.chris.tivi.calls
+package me.banes.chris.tivi.trakt.calls
 
 import com.uwetrottmann.trakt5.TraktV2
 import com.uwetrottmann.trakt5.entities.TrendingShow
@@ -28,6 +28,7 @@ import me.banes.chris.tivi.data.entities.TiviShow
 import me.banes.chris.tivi.data.entities.TrendingEntry
 import me.banes.chris.tivi.data.entities.TrendingListItem
 import me.banes.chris.tivi.extensions.toRxSingle
+import me.banes.chris.tivi.trakt.TraktShowFetcher
 import me.banes.chris.tivi.util.AppRxSchedulers
 import javax.inject.Inject
 
@@ -36,9 +37,9 @@ class TrendingCall @Inject constructor(
         showDao: TiviShowDao,
         trendingDao: TrendingDao,
         traktShowFetcher: TraktShowFetcher,
-        trakt: TraktV2,
+        private val trakt: TraktV2,
         schedulers: AppRxSchedulers
-) : PaginatedEntryCallImpl<TrendingShow, TrendingEntry, TrendingListItem, TrendingDao>(databaseTxRunner, showDao, trendingDao, trakt, schedulers, traktShowFetcher) {
+) : PaginatedEntryCallImpl<TrendingShow, TrendingEntry, TrendingListItem, TrendingDao>(databaseTxRunner, showDao, trendingDao, schedulers, traktShowFetcher) {
 
     override fun networkCall(page: Int): Single<List<TrendingShow>> {
         // We add one to the page since Trakt uses a 1-based index whereas we use 0-based
